app/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
cached-property==1.5.2
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==42.0.2
defusedxml==0.7.1
Django==3.2.23
django-celery-beat==2.2.1
django-celery-results==2.5.1
django-dockerizer==0.3.0
django-modeltranslation==0.18.11
django-timezone-field==4.2.3
djangorestframework==3.14.0
dnspython==2.5.0
exchangelib==4.7.2
idna==3.6
isodate==0.6.1
kombu==5.3.5
lxml==5.1.0
oauthlib==3.2.2
Pillow==10.1.0
prompt-toolkit==3.0.43
pycparser==2.21
Pygments==2.17.2
pyspnego==0.10.2
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
six==1.16.0
somepackage==1.2.3
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
uWSGI
